[#206] rename: Comment/Post LikeRepository 패키지 이동
