+++ src/main/java/com/konggogi/veganlife/post/repository/PostLikeRepository.java
@@ -1,4 +1,4 @@
-package com.konggogi.veganlife.like.repository;
+package com.konggogi.veganlife.post.repository;
 
 
 import com.konggogi.veganlife.like.domain.PostLike;
