+++ src/main/java/com/konggogi/veganlife/post/service/PostLikeQueryService.java
@@ -2,7 +2,7 @@ package com.konggogi.veganlife.post.service;
 
 
 import com.konggogi.veganlife.like.domain.PostLike;
-import com.konggogi.veganlife.like.repository.PostLikeRepository;
+import com.konggogi.veganlife.post.repository.PostLikeRepository;
 import java.util.Optional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
